Function sunrise_keyboard::ps2::get_event

pub fn get_event() -> &'static ReadableEvent

Get a ReadableEvent for the PS2 IRQ. Waiting on this event will wait until a keypress is detected. Note that once this event is triggered, it won't trigger again until read_key or try_read_key is called.