Struct sunrise_keyboard::Keyboard

struct Keyboard {
    writable_event: Option<WritableEvent>,
    readable_event: ReadableEvent,
    keys_queue: VecDeque<HidKeyboardState>,
}

Keyboard handling structure.

Fields
writable_event: Option<WritableEvent>

The event used to signal changes in the shared memory.

readable_event: ReadableEvent

The event returned to the client when requested via IPC.

keys_queue: VecDeque<HidKeyboardState>

The queue containing the keyboard state received by the driver.

Implementations

impl Keyboard

pub fn new() -> Result<Self, Error>

Create a new instance of Keyboard.

pub fn get_readable_event(&self) -> HandleRef<'static>

Get the readable update event of the Keyboard.

pub fn take_writable_event(&mut self) -> Option<WritableEvent>

Get the writeable update event of the Keyboard.

Note:

This consume the internal writable_event.q

pub fn handle_ps2_irq(&mut self) -> Option<()>

Handle a PS2 IRQ and push a new key state to the internal queue if needed.

pub fn read_keyboard_states(
&mut self,
states: &mut [HidKeyboardState]
) -> Result<u64, Error>

Get the last key states on the internal queue.